% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifr_colours.R
\name{ifr_colours}
\alias{ifr_colours}
\title{Loads IFR colour palette for less than six colours}
\usage{
ifr_colours()
}
\description{
This is an internal function to load IFR colour palette.
}
\details{
None.
}
\examples{
ifr_colours() +
}
