% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ifr_colours10.R
\name{ifr_colours10}
\alias{ifr_colours10}
\title{Loads IFR colour palette for seven to ten colours}
\usage{
ifr_colours10()
}
\description{
This is an internal function to load IFR colour palette.
}
\details{
None.
}
\examples{
ifr_colours10() +
}
