% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_colour_ifr.R
\name{scale_colour_ifr}
\alias{scale_colour_ifr}
\title{Apply IFR colour palette to ggplot object with six colours or less}
\usage{
scale_colour_ifr(...)
}
\description{
This function is added to a ggplot object to apply a custom IFR colour palette.
}
\details{
A colour palette based on the IFR logo that can be applied to plots with siz colours or less.
For plots with siz to ten colours, use \code{\link{scale_colour_ifr10}}.
}
\examples{
ggplot(mtcars, aes(as.factor(carb), fill = as.factor(carb))) +
  geom_bar() +
  scale_fill_ifr() +
  theme_ifr()
}
