% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_ifr_colours.R
\name{show_ifr_colours}
\alias{show_ifr_colours}
\title{Display the IFR colour palette}
\usage{
show_ifr_colours()
}
\description{
This function allows the user to view the colours and corresponding hex values used
within \code{\link{scale_colour_ifr}}.
}
\details{
This function can be used to display IFR colours and select individual colours for custom plots.
For example, when adding custom lines to a plot, IFR red (#a40c01) can be used instead of "red".
}
\examples{
show_ifr_colours() +
}
