% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_ifr.R
\name{theme_ifr}
\alias{theme_ifr}
\title{Apply IFR theme to ggplot2 object}
\usage{
theme_ifr(base_size = 16)
}
\arguments{
\item{base_size}{Adjust the base size of plot text.}
}
\description{
This function is added to a ggplot object to apply a custom IFR style.
}
\details{
This custom theme adjusts the standard ggplot theme to create report-ready plots.
Font sizing is appropriate for a plot approximately 6" square, or similar.
For larger or smaller plots, use base_size to adjust the text size as necessary to create
aesthetic plots with readable text. All theme adjustments can be over-ridden with a subsequent theme()
call.
}
\examples{
ggplot(mtcars, aes(mpg, disp)) +
  geom_point() +
  theme_ifr()
}
